fix: issues resolved with ci run

+++ scripts/installer/framework-generator.ts
@@ -54,8 +54,10 @@ const frameworkGenerator = ({
     shelljs.cp(reactMUIPackageJson, `${shelljs.pwd()}/package.json`);
     shelljs.exec(`cross-env INTERCEPTORS_FRAMEWORK_NAME=${interceptorFrameworkName} UTILS_FRAMEWORK_NAME=${utilsFrameworkName} COMPONENTS_FRAMEWORK_NAME=${frameworkName} UI_FRAMEWORK_NAME=${uiFrameworkName} npm install --force`);
     shelljs.exec(`cross-env INTERCEPTORS_FRAMEWORK_NAME=${interceptorFrameworkName} UTILS_FRAMEWORK_NAME=${utilsFrameworkName} COMPONENTS_FRAMEWORK_NAME=${frameworkName} UI_FRAMEWORK_NAME=${uiFrameworkName} npm run postinstall`);
-    shelljs.rm('-rf', `${shelljs.pwd()}/package.json`);
-    shelljs.mv(`${shelljs.pwd()}/package-original.json`, `${shelljs.pwd()}/package.json`);
+    if (!process.env.CI_RUN) {
+        shelljs.rm('-rf', `${shelljs.pwd()}/package.json`);
+        shelljs.mv(`${shelljs.pwd()}/package-original.json`, `${shelljs.pwd()}/package.json`);
+    }
 }
 
 module.exports = frameworkGenerator;
